gltf: Move stream-related functionality to stream.cpp
For now this includes both quantization bounds and actual stream
writing/compression functions.
