+++ CMakeLists.txt
@@ -33,11 +33,12 @@ set(SOURCES
 set(GLTF_SOURCES
     gltf/animation.cpp
     gltf/basistoktx.cpp
+    gltf/gltfpack.cpp
     gltf/material.cpp
     gltf/mesh.cpp
     gltf/parseobj.cpp
     gltf/parsegltf.cpp
-    gltf/gltfpack.cpp
+    gltf/stream.cpp
 )
 
 if(MSVC)
